Move token deletion logic to service layer
